requirements.txt
python-binance
requests
loguru